Add funding expiry to channel open
Routing nodes have an incentive to use low fees when opening channels to
ensure their activity is economically viable.

However, when a funding transaction takes too long to confirm, the fundee
may have forgotten the channel. In that case the funder is forced to
broadcast the first commit tx to get his funds back and then open a new
channel, which is costly.

We can avoid this issue by having an explicit commitment to a block before
which the funding tx will be confirmed. The fundee will keep the channel
around for that duration, and the funder needs to ensure the funding tx
confirms before that deadline.